test(design): add test cases
