--- packages/design/src/components/container/index.module.less
@@ -1,4 +1,0 @@
-.container-wrapper {
-    width: 100%;
-    height: 100%;
-}
